Use distant future OAuth 1 timestamps
Since the OAuth 1 timestamp is generated at page load time, if the
developer takes a while before clicking launch, or keeps the page open
and uses the launch button multiple times, the timestamp would expire
and the requests would start failing.

Fix this by setting the timestamp to one year in the future.

Note: I don't think this is really correct. The app that's being tested
_should_ probably refuse future timestamps (?). The correct solution
would be to have the form automatically set the timestamp to the current
time _when the user clicks launch_. But for now this hack seems to work
for our LTI app.